Fix conflict between expected changed files and git dirty files
For a Flutter project, specifying gradle_file_folder as "app" in the
lane config would lead to the following expected changed files
"app/build.gradle" whereas the git dirty file would actually be
"android/app/build.gradle".

This happens because the git root is not the android directory but the
flutter project folder which contains the android project in the
"android" directory.